Xilinx: ARM: MTD: fixing CFI for Numonyx CFI bug
commit  bb5bb3c3c0532e22f17a2c6e347cc2810dae1750 from
https://github.com/Xilinx/linux-xlnx.git

The M29EW devices seem to report the CFI information wrong when
it's in 8 bit mode.

There's an app note from Numonyx on this issue and there's a patch
in the open source, but it doesn't seem to be in the mainline.

This commit is based on that patch from Massimo Cirillo titled:
buffer size adjustment for M29EW Numonyx devices.

Signed-off-by: John Linn <john.linn@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>