mtd: nand: onfi don't WARN if we are in 16 bits mode
commit  2bab3191e25a35d97f3e50e7d7c24be34a686d85 from
https://github.com/Xilinx/linux-xlnx.git

Commit ff3206b2450499203532af2505a7f6f8413e92c0 ('mtd: nand: onfi need
to be probed in 8 bits mode') adds a WARN if the onfi probe is in 16
bits mode. This allows to detect driver that need to be fixed, but this
is a bit noisy¹. Transform the WARN in a pr_err.

¹ http://article.gmane.org/gmane.linux.ports.arm.omap/91317

Signed-off-by: Matthieu CASTET <matthieu.castet@parrot.com>
Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>