trafgen: xilinx: add axi traffic generator driver
commit  a7c9063abecaeb4700e72ab8b8fda16efd88b880 from
https://github.com/Xilinx/linux-xlnx.git

This is the driver for AXI Traffic Generator IP. The AXI
Traffic Generator IP is a core that stresses the AXI4
interconnect and other AXI4 peripherals in the system.
It generates a wide variety of AXI4 transactions based on
the core programming.

Architecture of the core is broadly separated into a master
and slave block, each of which contains the write block and
read block. Other support functions are provided by the
control registers and three internal RAMs - Master RAM,
Command RAM, Parameter RAM. The initialisation sequence
includes programming Command RAM with commands, data into
Master RAM (optional Parameter RAM programming) and then
enable master logic using control register interface.
This sequence generates traffic to cores connected in the
h/w design.

The driver for this IP is designed to be a module with
sysfs interface.  All the control registers and internal
RAMs can be accessed through sysfs interface.

NOTE: All the sysfs functions need to be documented
as per kernel-doc format.

This patch also adds support for Static Mode and Streaming Mode.
Use case: Streaming mode is used to stress the different Streaming
interfaces. Static mode is used to generate the required burst length
transactions.

In 64-bit support for ATG the CMDRAM is divided into two 4 KB regions,
one for reads and one for writes each region of CMDRAM can hold 256 commands.
And each command of command ram is 4 words or 128 bits wide.

Reads are issued to the master-read block AR channel from CMDRAM (0x000 to 0xFFF)
locations (up to 256 commands of 128 bits each). Writes are issued to the master-write
block AW channel from CMDRAM (0x1000 to 0x1FFF) locations (up to 256 commands of 128
bits each).

The 128 bits contain 32 bis of address. For supporting 64-bits
The ADDRRAM is added.

When the address width is configured in Vivado IDE is >32, the extended address capability
for the Master AXI4 interface in the AXI4 mode is available.

The Address RAM entries correspond to the MSB bits of address and are concatenated to
Bits[31:0] in CMDRAM.

The last_valid_idx has higher 16 bits for write and lower
16 bits for read. Mask the values according.

Allow disabling the errors.

Report the master done also in the error status

Add loop enable support to ATG.

Add the clock support for the axi traffic generator.

Fix  the offset for the  extended command ram

Signed-off-by: Srikanth Thokala <sthokal@xilinx.com>
Signed-off-by: Shubhrajyoti Datta <shubhraj@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>