MLK-11293: gpio: pca953x: add device_reset() call
commit  330d56ce1c1ace2fc96736f24b2ca15536531a9a from
https://source.codeaurora.org/external/imx/linux-imx.git

The pca953x type of devices, e.g. max7310, may have a reset which needs
to be handled to get the device start working.  Add a device_reset()
call for that, and defer the probe if the reset controller for that is
not ready yet.

Signed-off-by: Shawn Guo <shawn.guo@freescale.com>
Signed-off-by: Fugang Duan <B38611@freescale.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>