MLK-9777 cpufreq: imx6: fix the high bus count mismatch
commit  0c48693869f4051256a12f33173addc82a3e9bf2 from
https://source.codeaurora.org/external/imx/linux-imx.git

Normally, the system is booting up with higher cpufreq. In the
cpufreq set_target_index we will release the high bus mode if
the target cpu frequency is the lowest. It will release the high
bus mode and dcrease the high_bus_count.This will lead to a wrong
release of high bus mode. So, in the cpufreq_init function, if the
original frequency is not the lowest, we need request high busfreq.

Signed-off-by: Bai Ping <b51503@freescale.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>