ARM: axxia: Add cluster and L2 power off in hotplug
Adds a choice to hotplug. There are two ways to power down the cpu,
either to a low power mode or completely off. If the power off mode
is selected then the a new option is available to power off the
L2 cache as well.

This code was originally written by Charlie Paul
<Charlie.Paul@windriver.com> and John Jacques <john.jacques@intel.com>.
Only minor modifications performed by the author of this patch.

Signed-off-by: Charlie Paul <Charlie.Paul@windriver.com>
Signed-off-by: John Jacques <john.jacques@intel.com>
Signed-off-by: Magnus Karlsson <magnus.karlsson@intel.com>
[Quanyang:Original patch taken from
git://git.yoctoproject.org/linux-yocto-4.12 standard/axxia/base]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>