misc: mtc: added MTC driver code for Axxia5600
MTC driver is an AXXIA driver for the Master Test Controller interface.
This patch adds MTC driver support for AXXIA 56xx Emulator.
Some of the features are yet to be tested when the board is available.
Basic test that accesses the driver runs without errors.

Signed-off-by: Sreedevi Joshi <sreedevi.joshi@intel.com>
[Quanyang:Original patch taken from
git://git.yoctoproject.org/linux-yocto-4.12 standard/axxia/base]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>