spi-bcm2835: Support pin groups other than 7-11
commit ea7ff6bd869087be46f2d8f8ac59e85866fac12e from
https://github.com/raspberrypi/linux.git

The spi-bcm2835 driver automatically uses GPIO chip-selects due to
some unreliability of the native ones. In doing so it chooses the
same pins as the native chip-selects would use, but the existing
code always uses pins 7 and 8, wherever the SPI function is mapped.

Search the pinctrl group assigned to the driver for pins that
correspond to native chip-selects, and use those for GPIO chip-
selects.

Signed-off-by: Phil Elwell <phil@raspberrypi.org>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>