MLK-11561-2 regulator: anatop: set default voltage selector for pcie
commit  818b3b9001809377494ace2711a91e13dfc07267 from
https://source.codeaurora.org/external/imx/linux-imx.git

The code reads the default voltage selector from its register.
If the default voltage selector is 0 which results in faulty
behaviour of this regulator driver.

This patch sets a default voltage selector for vddpcie-phy if
it is not set in the register.

Signed-off-by: Richard Zhu <Richard.Zhu@freescale.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>