pwm: rcar: Add a judgment of the period out of range
commit fb813e240cecd77ee4133de5990343c29a0c58f7 from
git://git.kernel.org/pub/scm/linux/kernel/git/horms/renesas-bsp.git

Current code does not cover the case that user inputs invalid
value of period.
This patch adds a judgment to output an error when an input period
value is out of the range. As description in hardware manual the
minimum value of period is not smaller than 2 clock cycles.

Signed-off-by: Ryo Kodama <ryo.kodama.vz@renesas.com>
Signed-off-by: Van Do <van.do.xw@renesas.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>