MLK-12072 thermal: imx: enable tempmon finish bit check on imx7d TO1.1
commit  5ccc424b81850fd55aea7e2c9ff8c254d005f733 from
https://source.codeaurora.org/external/imx/linux-imx.git

On i.MX7D TO1.0, the finish bit in tempmon module used for verify
the temp value is broken, so it can NOT be used for checking the temp
value. On TO1.1, this issue has been fixed, so we can use this bit
to verify if the temp value is valid.

Signed-off-by: Bai Ping <ping.bai@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>