MLK-12103 media: capture: ov5640_mipi: add more delay for camera to stay stable
commit  fcc51f4eb0e38b631b1380b114c434dde7097f65 from
https://source.codeaurora.org/external/imx/linux-imx.git

For power saving, the camera is set ON/OFF by 'PWDN' pin.
However, the DQBUF timeout problems have been observed occasionally when
do the resolution switch for the camera. The reason could be that the power
for the camera is set to off then on when do the switch, but the delay time
is not enough for camera to go stable. The patch sets it to an empirical
value and it has passed over 3-day stress test.

Signed-off-by: Robby Cai <robby.cai@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>