zynq: xadc: limit pcap clock frequency value
commit  bfe5046c821650d656b783d5d4adf876d0e70a38 from
https://github.com/Xilinx/linux-xlnx.git

This patch limits the xadc pcap clock frequency value to be less than
200MHz.

This workaround is required to avoid the below issue:

When zynq is booted at higher frequency values, pcap crosses the maximum
limit of 200MHz(Fmax) as it is derived from IOPLL.
If this limit is crossed it is required to alter the WEDGE and REDGE
bits of XADC_CFG register to make timings better in the interface.
So to avoid alteration of these bits every time, the pcap value should
not cross the Fmax limit.

Signed-off-by: Ranjit Waghmode <ranjit.waghmode@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>