MLK-12202: ASoC: fsl_asrc_m2m: fix dma timeout after suspend/resume
commit  3c65f04560133f5d0690ee46b636b4ac9e3e31e8 from
https://source.codeaurora.org/external/imx/linux-imx.git

In suspend function, the complete will be set to done in callback.
After resuming, the convert will not spend time to wait the complete.
which is a wrong complete.
So in suspend function, the complete need to be reinited for next convert.

Signed-off-by: Shengjiu Wang <shengjiu.wang@freescale.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>