MLK-12244 ASoC: fsl: Set the sound card owner field
commit  b08c460cfa2ca079714dc574c591ad63b385c2a7 from
https://source.codeaurora.org/external/imx/linux-imx.git

When load sound card, the pulseaudio will using the sound card to
playback and record. It may be cause a kernel crash when the sound
card is unloaded while the playback/record is active

After setting the sound card owner field, when pulseaudio is running,
the sound card module ref-count will not be 0 and the sound card will
not be unloaded. So you should stop the pulseaudio when you want to
unload the sound card.

Signed-off-by: Zidan Wang <zidan.wang@freescale.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>