MLK-12433: arm: dts: imx: disable tempmon device when cpufreq is disabled
commit  2b78e693ffaec7c8640f7b1d7c5969ab072bef2b from
https://source.codeaurora.org/external/imx/linux-imx.git

In the lpddr3-arm2-m4 dts, the I2C1 is disabled, so PMIC is disabled,
the cpufreq is not support. As thermal driver is depended on cpufreq
driver, if cpufreq is not support, the tempmon device can be disabled.

Signed-off-by: Bai Ping <ping.bai@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>