MLK-12794: ASoC: fsl_asrc: fix underrun issue when convert 192k to 96kHz.
commit  d055634c39d12e78730657bfbc03d3a2721848e1 from
https://source.codeaurora.org/external/imx/linux-imx.git

The maximum divider of asrc clock is 1024, but there is no judgement
for this limitaion in driver, which may cause the divider setting not
correct.

When IDEAL_RATIO_RATE 200kHZ, the cost time of conversion from 192kHz
to 96kHz is 24ms every 1024 sample, but these sample's playback time
is 1024/96=11ms, so there will be underrun. So need to enlarge this RATE.

Signed-off-by: Shengjiu Wang <shengjiu.wang@freescale.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>