spi: sh-msiof: Add sleep before master transfer for test
commit 9a9d0e5d25e5ab8c7ee2d3a87c4af1353f071dbc from
git://git.kernel.org/pub/scm/linux/kernel/git/horms/renesas-bsp.git

This patch is for debug of transfer between master and slave.
Since the slave needs to complete a preparation in data transfer
before the master working, the sleep wait is put before
the data transfer of the master.

Signed-off-by: Hiromitsu Yamasaki <hiromitsu.yamasaki.ym@renesas.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>