axxia: Updates to the NEMAC Network Interface Driver
Make the transfer length a multiple of 64 bytes.  The
number of bytes transferred will still be limited to
the pdulength.  Hardware requires the transfer length
to be a multiple of 64.

Add a memory barrier after updating the descriptor in
memory but before incrementing the head pointer to
start packet transmission.

Signed-off-by: John Jacques <john.jacques@intel.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>
[Quanyang:Original patch taken from
git://git.yoctoproject.org/linux-yocto-4.12 standard/axxia/base]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>