watchdog: bcm2835: Support setting reboot partition
commit 3bb6ebcaad1cbbbb6e9540d4ddce76540aa26ac6 from
https://github.com/raspberrypi/linux.git

The Raspberry Pi firmware looks at the RSTS register to know which
partition to boot from. The reboot syscall command
LINUX_REBOOT_CMD_RESTART2 supports passing in a string argument.

Add support for passing in a partition number 0..63 to boot from.
Partition 63 is a special partiton indicating halt.
If the partition doesn't exist, the firmware falls back to partition 0.

Signed-off-by: Noralf Trønnes <noralf@tronnes.org>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>