MLK-13609: ASoC: fsl_sai: fix for synchronize mode
commit  e91a502b50161f6449a3d8a44b02b13f0d619c31 from
https://source.codeaurora.org/external/imx/linux-imx.git

TX synchronous with receiver: the RMR should not be changed and
the RCSR.RE should be set in playback.
RX synchronous with transmitter: the TMR should not be changed and
the TCSR.TE should be set in recording.

Signed-off-by: Shengjiu Wang <shengjiu.wang@freescale.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>