PCI: pcie-rcar: HACK: Add the initialization of PCIe link in resume_noirq
commit d3468b2d62a00e51bea689d8ed004a5cf4b6e7e4 from
git://git.kernel.org/pub/scm/linux/kernel/git/horms/renesas-bsp.git

In PM noirq resume, it calls the read function of pci_ops from
pci framework. Thus, it needs to initialize the rcar-pcie register
in case there is calling in noirq resume. But it must be revised
the structure of rcar-pcie before this is fixed.

Therefore, it will make interim fix in this patch until changing
the structure of rcar-pcie. Also it fixes msleep together since
it can't be used msleep in noirq phase.

Signed-off-by: Kazufumi Ikeda <kaz-ikeda@xc.jp.nec.com>
Signed-off-by: Gaku Inami <gaku.inami.xw@bp.renesas.com>
Signed-off-by: Tho Vu <tho.vu.wh@rvc.renesas.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>