remoteproc: check if remote is running the fw before reload fw
commit  52dba1ddcad0cc2f73fd0d4be0f97d5eea6d6f49 from
https://github.com/Xilinx/linux-xlnx.git

Add a function rproc_is_running_fw() to check if it needs to start the
remote betfore it loads firmware and tries to start the remote.

In some cases, when the master restarts, the remote is already running the
expected firmware.
In this case, we don't need to start the remote when the remoteproc driver
starts.

Signed-off-by: Wendy Liang <jliang@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>