remoteproc: check firmware checksum in rproc_is_running_fw()
commit  09acc0693c2f4c5ecf86a591c72d95ef8632ee52 from
https://github.com/Xilinx/linux-xlnx.git

In rproc_is_running_fw() function, set the firmware checksum in resource
in the resource table and compared it with the one in the loaded resource
table.

If the firmware checksum in the loaded resource table matches the one
calculated with the firmware, the remote is running with the expected
firmware.

Signed-off-by: Wendy Liang <jliang@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>