remoteproc: virtio: kick in reset virtio
commit  70c5ad225b46e876c27f625df7cb2060db4a97b2 from
https://github.com/Xilinx/linux-xlnx.git

Kick the remote in virtio reset.
In order to be backward compatible, only set the vdev
status to NEEDS_RESET and kick the remote if the remote is
RUNNING_INDEPENDENT, that is only if the remote is already
running.

Signed-off-by: Wendy Liang <jliang@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>