MLK-13914-04 cpufreq: imx7ulp: Add cpufreq driver on i.mx7ulp
commit  ad056614869e2e267658515fae8a373311cc779c from
https://source.codeaurora.org/external/imx/linux-imx.git

According to the i.MX7ULP RM & datasheet, when A7 is in RUN mode,
the max CPU frequency it can run at is 500MHz, if we want to increase
the A7 CPU frequency to 800MHz, we need to change the A7 from RUN mode
to HSRUN mode. And only when A7 is in RUN mode, we can enter STOP mode
and VLLS mode.

Signed-off-by: Bai Ping <ping.bai@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>