MLK-14254: ASoC: imx_pcm_rpmsg: fix cmd dropped by work queue
commit  052367ab2334b4f6ac09f97ad0a002006df1fbe1 from
https://source.codeaurora.org/external/imx/linux-imx.git

The test case is to playback a bitstream, then repeat ctrl+z and fg,
several times later, the playback is failed to continue.

The reason is if the work is pending in work queue, send second time
of this work, the second work is dropped by work queue. so use one
work for one cmd is not fit for audio case. use a work loop for audio
cmd to fix this issue.

Signed-off-by: Shengjiu Wang <shengjiu.wang@freescale.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>