drivers: dwc3: Add U3 suspend disable quirk for silicon versions less than v4
commit  9b54b6b43f69a0aadc64bbc4ddfd2a7db4397e6d from
https://github.com/Xilinx/linux-xlnx.git

The logic for updating dwc->revision flag got corrected in 4.9 kernel,
which sets USB3PIPECTL.SUSPENDENABLE bit for dwc3 core versions >= 2.90A.
Because of this bit being set we are seeing failure while configuring
endpoints and below error is seen

"Failure to configure ep0out"

When this bit is set the control enables autosuspended of phy. As a part
of this feature controller automatically disables the phy clocking when
it is idle. This issue is observed with silicon revisions < 4.0.

This patch adds "snps,dis_u3_susphy_quirk" into platform node if silicon
revision is found to be less than 4.0

This patch is based on the fix provided in "85e618a1be2b2092318178d1d66b"

Signed-off-by: Anurag Kumar Vulisha <anuragku@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
fs
Signed-off-by: Meng Li <Meng.Li@windriver.com>