phy: dp83867: macb: SW workaround for link instability
commit  1e67e4abd9613f074f1a8f27e7e008e37f48367c from
https://github.com/Xilinx/linux-xlnx.git

RX_CTRL of DP83867 has to be strapped to MODE 3 or 4 according to
TI datasheet. If this is not done link instability may be observed
with some partners. This was observed on ZCU102 at 100Mbps and
also previously with on SGMII DC when connected to switches.
From communication with TI, we learnt that the following can be
used as a SW workaround to solve these issues.
Added this workaround with a DT property to make it configurable.

Signed-off-by: Harini Katakam <harinik@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>