Add support for the AudioInjector.net Octo sound card
commit 2f780d6169cbeb08e93783a55e395dfcb5303ae6 from
https://github.com/raspberrypi/linux.git

AudioInjector Octo: sample rates, regulators, reset

This patch adds new sample rates to the Audioinjector Octo sound card. The
new supported rates are (in kHz) :
96, 48, 32, 24, 16, 8, 88.2, 44.1, 29.4, 22.05, 14.7

Reference the bcm270x DT regulators in the overlay.

This patch adds a reset GPIO for the AudioInjector.net octo sound card.

Audioinjector octo : Make the playback and capture symmetric

This patch ensures that the sample rate and channel count of the audioinjector
octo sound card are symmetric.
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>