net: ethernet: axienet: Fix bug in the rxreject interrupt handling
commit  91a905c2748c6df03c1abc7e83e158ad237381c6 from
https://github.com/Xilinx/linux-xlnx.git

Currenlty driver is doing a reset of the device
when we got a rxreject interrupt which is not proper,
when we got a rxreject interrupt it means ethernet mac
won't have enough room to store the recv packet.
In such scenario we should increment the device dropped
Packet statstics instead of resetting the device.

This patch fixes the same.

Signed-off-by: Kedareswara rao Appana <appanad@xilinx.com>
Acked-by: Anirudha Sarangi <anirudh@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>