MLK-14454: video: mipi_dsi_northwest: Fix leaking a reset_control
commit  4022b22bbdc09ac1ec650c7eb5bb5d2599ce2ec4 from
https://source.codeaurora.org/external/imx/linux-imx.git

The mipi_dsi_disp_init function takes a reference to a reset_control
during initialization and does not release it.

This becomes a problem on 4.9 because reset_control_get has been split
into exclusive mode(default) and shared(which need to be marked).
Leaking a reference counts looks like attempting to fetch a second
reference to a controller and this WARNs and returns -EBUSY.

Fix by releasing it at the end of the function.

Signed-off-by: Leonard Crestez <leonard.crestez@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>