arm64: dts: r8a7795: Add cpu capacity-dmips-mhz
commit 1f213836d1de2123a7fe4deb9eb5829a72c81fda from
git://git.kernel.org/pub/scm/linux/kernel/git/horms/renesas-bsp.git

Set the capacity-dmips-mhz for R-CAR H3, that is based on dhrystone.

Expected cpu capacity:
Cortex-A57: 1024, Cortex-A53: 432

Note:
The cpu capacity in Cortex-A53 side is scaled with max cpufreq
in current implementation.

540 * 1.2[GHz]/1.5[GHz] = 432

Therefore, it needs to set "540" to capacity-dmips-mhz in dt
if it uses cpu capacity with "432" on Cortex-A53.

Signed-off-by: Gaku Inami <gaku.inami.xw@bp.renesas.com>
Signed-off-by: Takeshi Kihara <takeshi.kihara.df@renesas.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>