arm64: dts: r8a7796: Add cpu capacity-dmips-mhz
commit 6cb8f68eb1dece1ac119beea58bb09f8ea2a9313 from
git://git.kernel.org/pub/scm/linux/kernel/git/horms/renesas-bsp.git

Set the capacity-dmips-mhz for R-CAR M3, that is based on dhrystone.

Expected cpu capacity:
Cortex-A57: 1024, Cortex-A53: 434

Note:
The cpu capacity in Cortex-A53 side is scaled with max cpufreq
in current implementation.

543 * 1.2[GHz]/1.5[GHz] = 434

Therefore, it needs to set "543" to capacity-dmips-mhz in dt
if it uses cpu capacity with "434" on Cortex-A53.

Signed-off-by: Gaku Inami <gaku.inami.xw@bp.renesas.com>
Signed-off-by: Takeshi Kihara <takeshi.kihara.df@renesas.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>