usb: dwc3: core: Enable CCI support in dwc3 core driver
commit  98379e0ad3dc2f68fa864eece20d728a106e2f50 from
https://github.com/Xilinx/linux-xlnx.git

This patch adds CCI support in DWC3 driver when CCI is enabled in the
design. There are couple of registers which need to be modified for
making CCI enabled transfers in USB. This patch add the support for
the same.

Signed-off-by: Manish Narani <mnarani@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>