MLK-14644 PCI: imx: toggle the perst after the release of rc's reset
commit  864eb40ead7877d56b8fc9d36bac5983b8cc0c8b from
https://source.codeaurora.org/external/imx/linux-imx.git

imxcommunity reports that Intel Wireless 3160 doesn't
work if the perst# is toggled before the release of rc
controller's reset.

Specification doesn't specify the exactly sequence of
the toggle of the perst# and the release of the rc
controller reset. Just find the following description
in the chapter 4.2.2 of the PCI Local Bus Specification.

"The system must guarantee that the bus remains in the idle
state for a minimum time delay following the deassertion
of RST# to a device before the system will permit the first
assertion of FRAME#."

Toggle the perst# after the release of rc's reset to
fix the issue.

Signed-off-by: Richard Zhu <hongxing.zhu@nxp.com>
(cherry picked from commit ba75c5e4795ab36d7765fe3ac03f0c4320828c78)
Signed-off-by: Meng Li <Meng.Li@windriver.com>