MLK-14536: ASoC: wm8960: Fix playback in CPU DAI master mode
commit  174c20dae01120e3b0ccbec1bf51c678e5952486 from
https://source.codeaurora.org/external/imx/linux-imx.git

With the current rates for MCLK is not possible to derive bitclk
for all files in S20_3LE format and also for files with S24_LE sampled
at 48000Hz.

In order to fix this, we need to find a better MCLK value. We did this
in two steps:
	1) Use params_physical_width to get rid of S20_3LE burden.
	2) Brute force into all available rates which can pass fsl_sai_set_bclk
	   algorithm.

Thus we found 36864000 to be the smallest acceptable rate for MCLK.

Reviewed-by: Mihai Serban <mihai.serban@nxp.com>
Suggested-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Signed-off-by: Daniel Baluta <daniel.baluta@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>