soc: fsl: dpio: change CENA regs to be cacheable
commit 81ef0840c90ad2b414b2025b43d485b1198eb385 from
http://source.codeaurora.org/external/qoriq/qoriq-components/linux.git

Change cache enabled regsiter accessed to be cacheable
plus non-shareable to meet the performance requirement.
QMan's CENA region contains registers and structures that
are 64byte in size and are inteneded to be accessed using a
single 64 byte bus transaction, therefore this portal
memory should be configured as cache-enabled. Also because
the write allocate stash transcations of QBMan should be
issued as cachable and non-coherent(non-sharable), we
need to configure this region to be non-shareable.

Signed-off-by: Haiying Wang <Haiying.Wang@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>