MLK-14935: ASoC: fsl_sai: Fix mixing initialization data with actual audio samples
commit  171201390019a79851790e452f0f5813c1c70312 from
https://source.codeaurora.org/external/imx/linux-imx.git

When starting a playback the initialization data used to reduce underruns
was send to the transmit data register after the DMA requests were enabled.
This patch moves the initialization phase before enabling the DMA so the
data is transmitted in correct order.

Signed-off-by: Mihai Serban <mihai.serban@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>