drivers/usb/core: fix over-current race condition
There exists a race condition when an over-current fault is cleared,
between the event processing in the hub_event function and the setting
of the event bits in the hub_irq function. When the over-current is
cleared the hub_event call tries to read the event bits, which havent
been set by the hub_irq.  In this patch when the over-current is
cleared the hub_event will check the event bits and if they are not
set it will pause and check again. It will sync the event bits.

Signed-off-by: Charlie Paul <Charlie.Paul@windriver.com>
Signed-off-by: John Jacques <john.jacques@intel.com>
[Quanyang:Original patch taken from
git://git.yoctoproject.org/linux-yocto-4.12 standard/axxia/base]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>