crypto: caam - add support for MOVEB command
CHAs of SEC work natively in BE mode. When moving
data to the alignment blocks, swapping is needed
for LE platforms. This is done by means of the MOVEB
command. This patch adds support
to DCL for this command.

Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
Signed-off-by: Radu Alexe <radu.alexe@nxp.com>
(cherry picked from commit e1277a1ebbb1a90833757d59d37c2e1487770071)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>