kexec: adjust dtb_mem address to adapt to axm55xx's weird entry
When kexec load secondary kernel, it will calculate dtb_mem according to
the formula as below:
dtb_mem = kernel_entry + image_size * 5
But axm55xx's entry is not the default value 0x8000 but 0x408000, this will
result that when zImage as the secondary kernel decompresses,
kernel_entry + kernel_size + zImage_size > dtb_mem, and data of dtb file in
dtb_mem will be overlapped by zImage. So add 4M to dtb_mem to avoid this.

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>