net: ethernet: xilinx: Fix ping not working after ifdown and ifup
commit  8e73144c3daaf87fc90dab474065da16fada8415 from
https://github.com/Xilinx/linux-xlnx.git

With the existing driver if user does a ifdown and ifup
of the ethernet interface data integrity is not working.
We are stoping the ndev in the close/down but not starting
the ndev in the open/up.

This patch fixes this issue by starting the ndev in the open/up().

Signed-off-by: Kedareswara rao Appana <appanad@xilinx.com>
Acked-by: Harini Katakam <harinik@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>