MLK-15083 watchdog: imx2_wdt: fallback to timeout reset if explicit reset fails
commit  fd54e52c59984994db181389ae91c7f3030a9641 from
https://source.codeaurora.org/external/imx/linux-imx.git

If explicit reset fails fallback using the watchdog timeout. We
already have set the timeout counter to 0, but we might need to ping
the watchdog to load the new timeout, if a previous watchdog timeout
value has already been set.

We also decrease the time we spend waiting, to give a chance to log
that the explicit reset failed and that we fallback to watchdog
timeout reset.

Signed-off-by: Octavian Purdila <octavian.purdila@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>