MA-9807: Fix ecb(aes) use without an IV
commit  8da66528d2da766e1a82728cc7df3d2bdc570267 from
https://source.codeaurora.org/external/imx/linux-imx.git

CAAM aes modes share descriptors, because of this CAAM requires an IV
for ECB. ECB does not need an IV and users do not have to pass valid
IV vectors. To allow correct usage with minimum impact to the driver a
zero IV is provided by the driver for ECB operations that need it.

Signed-off-by: Radu Solea <radu.solea@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>