net: ethernet: xilinx: Fix race condition in the tx path
commit  3e7e305649d8c1b4849edf542a7697a92035c788 from
https://github.com/Xilinx/linux-xlnx.git

In the _xmit driver is checking for the buffer descriptor
availabilty, If the buffer descriptors are not available
it is stopping all the device queues,
It should stop only that particular queue as
other queues still may have free buffer descriptors
for servicing the data.

This patch fixes this issue.

Signed-off-by: Kedareswara rao Appana <appanad@xilinx.com>
Acked-by: Harini Katakam <harinik@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>