media: xilinx: Ensure Video Framebuffer selected by Kconfig
commit  c6271bb2253ed4d694065ae0daa9525e9a646636 from
https://github.com/Xilinx/linux-xlnx.git

The Xilinx Video Linux driver is a client of the Video Framebuffer DMA
driver. Because the Framebuffer driver exports symbols used by this driver,
it is important that both are built in the same manner: as modules or
as statically included drivers.

Signed-off-by: Jeffrey Mouroux <jmouroux@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>