MLK-16052-2 PCI: imx: fix the pcieb link down issue
commit  45ab6ee2a29bb881a51899aebe604a957104a57a from
https://source.codeaurora.org/external/imx/linux-imx.git

The calibration value of PCIEA is mandatory required
by PCIEB on iMX8QM.

The RSTs of the PCIEA would be cleared when the PDs
of PCIEA are turned off.
The calibration value of PCIEA would be lost when the
RSTs of PCIEA are cleared.
So, the RSTs of PCIEA should be asserted when enable
the PCIEB port. Otherwise, PCIEB wouldn't be functional.

Signed-off-by: Richard Zhu <hongxing.zhu@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>