MLK-16108 PCI: imx: turn on pd for imx8mq pcie
commit  d2cf1e1db1ce07f6b99bad6c622bc8216c572e59 from
https://source.codeaurora.org/external/imx/linux-imx.git

Root cause:
Poewr domain of the PCIEs are turned off, and
not turned on properly in previous ATF.

The PDs of PCIE1/2 have the dependency.
Both of the PDs should be operated at same time.
This issue is gone after update the PDs operations
in ATF.
In order to make sure that the PDs are turned on,
Turn power domain for imx8mq pcie explicitly in
driver.

Signed-off-by: Richard Zhu <hongxing.zhu@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>