MLK-16203 enable runtime pm of i2c temporary when do system suspend
commit  bd904a8045b3aeae079d90eecfdad6b3a848b4cf from
https://source.codeaurora.org/external/imx/linux-imx.git

When we do system suspend, the runtime pm will be disabled, but we need
to control the PMIC to power on/off the regulator, if the runtime pm is
disabled, if will failed to request runtime wakeup. So data transfer will
failed.

Signed-off-by: Bai Ping <ping.bai@nxp.com>
Signed-off-by: Gao Pan <pandy.gao@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>